Refactor "geometry" formulas
