+++ cfspopcon/formulas/__init__.py
@@ -1,0 +1,6 @@
+"""Formulas used for the POPCON analysis."""
+from . import geometry
+
+__all__ = [
+    "geometry"
+]
