+++ cfspopcon/formulas/geometry/__init__.py
@@ -1,0 +1,6 @@
+"""Formulas to calculate terms related to the plasma geometry, such as the volume inside the last-closed-flux-surface."""
+from . import analytical
+
+__all__ = [
+    "analytical"
+]
