+++ example_cases/SPARC_PRD/input.yaml
@@ -4,7 +4,9 @@ algorithms:
   # and temperature, which defines a stored energy stored_energy. From this,
   # we can compute P_in from stored_energy and a tau_e scaling.
   - read_atomic_data
-  - calc_geometry
+  - calc_geometry_from_ratios
+  - calc_plasma_volume
+  - calc_plasma_surface_area
   - calc_q_star_from_plasma_current
   - calc_fuel_average_mass_number
   - calc_average_ion_temp
